chore: revert layout changes to report section, add api usage to database page
