+++ studio/pages/project/[ref]/reports/index.tsx
@@ -1,35 +1,81 @@
-import { useState } from 'react'
+import { useEffect, useState } from 'react'
 import { observer } from 'mobx-react-lite'
 import { useRouter } from 'next/router'
-
 import { NextPageWithLayout } from 'types'
-import { useFlag, useStore } from 'hooks'
+import { useStore } from 'hooks'
+import { post } from 'lib/common/fetch'
+import { API_URL, PROJECT_STATUS } from 'lib/constants'
+import { useProjectContentStore } from 'stores/projectContentStore'
+import Loading from 'components/ui/Loading'
 import { ProjectLayoutWithAuth } from 'components/layouts'
-import PresetReport from 'components/interfaces/Reports/PresetReport'
-import { Presets } from 'components/interfaces/Reports/Reports.types'
-import ReportsLayout from 'components/layouts/ReportsLayout/ReportsLayout'
-import { DashboardReportPage } from './dashboard'
+import ProductEmptyState from 'components/to-be-cleaned/ProductEmptyState'
+import { createReport } from 'components/to-be-cleaned/Reports/Reports.utils'
 
-export const ReportsOverviewPage: NextPageWithLayout = () => {
+export const UserReportPage: NextPageWithLayout = () => {
+  const [loading, setLoading] = useState(true)
+
   const router = useRouter()
   const { ref } = router.query
-  const reportsOverview = useFlag('reportsOverview')
 
-  const Layout = reportsOverview ? ReportsLayout : ProjectLayoutWithAuth
+  const { ui } = useStore()
+  const project = ui.selectedProject
+
+  // const canCreateReport = checkPermissions(PermissionAction.CREATE, 'user_content', {
+  //   resource: { type: 'report' },
+  // })
+  const contentStore = useProjectContentStore(ref)
+
+  useEffect(() => {
+    if (project && project.status === PROJECT_STATUS.INACTIVE) {
+      post(`${API_URL}/projects/${ref}/restore`, {})
+    }
+  }, [project])
+
+  async function loadReports() {
+    await contentStore.load()
+    const reports = contentStore.reports()
+
+    if (reports.length >= 1) {
+      router.push(`/project/${ref}/reports/${reports[0].id}`)
+    } else {
+      setLoading(false)
+    }
+  }
+
+  useEffect(() => {
+    loadReports()
+  }, [ref])
 
   return (
-    <Layout>
-      {reportsOverview ? (
-        <PresetReport preset={Presets.OVERVIEW} projectRef={ref as string} />
+    <div className="mx-auto my-16 w-full max-w-7xl flex-grow space-y-16">
+      {loading ? (
+        <Loading />
       ) : (
-        <DashboardReportPage />
+        <ProductEmptyState
+          title="Reports"
+          ctaButtonLabel="Create report"
+          onClickCta={() => {
+            try {
+              createReport({ router })
+            } catch (error: any) {
+              ui.setNotification({
+                category: 'error',
+                message: `Failed to create report: ${error.message}`,
+              })
+            }
+          }}
+        >
+          <p className="text-scale-1100 text-sm">Create custom reports for your projects.</p>
+          <p className="text-scale-1100 text-sm">
+            Get a high level overview of your network traffic, user actions, and infrastructure
+            health.
+          </p>
+        </ProductEmptyState>
       )}
-    </Layout>
+    </div>
   )
 }
 
-// TODO: uncomment when reportsOverview flag is removed
-// hooks do not work with next.js .getLayout
-// ReportsOverviewPage.getLayout = (page) => <ReportsLayout>{page}</ReportsLayout>
+UserReportPage.getLayout = (page) => <ProjectLayoutWithAuth>{page}</ProjectLayoutWithAuth>
 
-export default observer(ReportsOverviewPage)
+export default observer(UserReportPage)
